### 2021-05-05 #3245 @jamesrhester
Can confirm that https://whedon.theoj.org/ compiles the paper from this repository if `joss_paper` is provided as the branch. Perhaps the version number causes a problem?
