### 2021-05-05 #3245 @kthyng
@openjournals/dev The JOSS paper is located by itself at branch `joss_paper`, but I can't compile it from there. Any ideas?
