### 2021-05-05 #3245 @arfon
> @openjournals/dev The JOSS paper is located by itself at branch joss_paper, but I can't compile it from there. Any ideas?

@kthyng - looks like you didn't quite have the syntax of the command correct :-)
