### 2021-06-14 #3245 @danielskatz
Hi @sbenthall  - the best option at this point is for you to sign yourself up, by doing what I am going to do below
