### 2021-06-14 #3245 @danielskatz
you can do this yourself, or you can also do `@whedon assign me as editor`
