### 2021-06-18 a118d10
Minor edits to README
