### 2021-07-15 #3497 @whedon
[REVIEW]: Lerche: Generating data file processors in Julia from EBNF grammars
**Submitting author:** @jamesrhester (<a href="http://orcid.org/0000-0002-2004-8672">James Hester</a>)
**Repository:** <a href="https://github.com/jamesrhester/Lerche.jl" target ="_blank">https://github.com/jamesrhester/Lerche.jl</a>
**Version:** 0.5.1
