### 2021-07-19 #3497 @sbenthall
@jamesrhester any ideas as to why there is a PDF compilation error? We'll need to fix this to enable to reviewers.
