### 2021-07-21 jamesrhester/Lerche.jl#18 @ziotom78
«Stable docs» badge does not work
Clicking on the «Docs | stable» badge in the README triggers a 404 error. («Docs | devel» works fine.)

**Disclaimer**: this issue was opened following the review happening here: https://github.com/openjournals/joss-reviews/issues/3497.
