### 2021-07-21 jamesrhester/Lerche.jl#19 @ziotom78
Undefined behavior when errors in the parser are found
The documentation does not mentions what happens when there is an error in the source text passed to the parser. It seems that, similarly to Lark, exceptions are thrown, and these exceptions share the same name as Lark's (`UnexpectedInput`, `UnexpectedCharacters`, `UnexpectedToken`):

```
julia> j = Lerche.parse(json_parser, raw"{\"key\": [\"item0\", \"item1\", 3.14]]")
Unexpected input: UnexpectedToken(Token(RSQB, ]), ["RBRACE", "COMMA"], 1, 33, nothing, 11, 33, nothing, false)
ERROR: Unexpected token ] at line 1, column 33.
Expected one of: RBRACE, COMMA

Stacktrace:
…
```

Although Lark's documentation lacks an organic presentation of how error are handled (but it does document `UnexpectedToken` and `UnexpectedCharacters` in the documentation for [`UnexpectedInput`](https://lark-parser.readthedocs.io/en/latest/classes.html#unexpectedinput)), I believe that the usage of exceptions to signal errors and their type should be mentioned somewhere in Lerche's documentation, as this is an important piece of information that helps to ensure robustness in codes.

**Disclaimer**: this issue was opened following the review happening here: https://github.com/openjournals/joss-reviews/issues/3497.
