### 2021-07-22 jamesrhester/Lerche.jl#19 @erezsh
@ziotom78 If you don't mind me asking, what do you mean by an organic presentation of how errors are handled?

Anyway, I improved the documentation on exceptions a little. Thanks for pointing that out. (https://lark-parser.readthedocs.io/en/latest/classes.html#lark.Lark.parse)
