### 2021-07-22 jamesrhester/Lerche.jl#19 @ziotom78
> @ziotom78 If you don't mind me asking, what do you mean by an organic presentation of how errors are handled?

No worry! I meant that a reader would like to find something better than what's inside Lark's documentation. I had to dig a bit to discover how errors are handled, and I managed to find it because I tested with some code that `UnexpectedToken` was raised whenever a token was in the wrong place and was then able to search for this keyword in the docs. It would have been far better to have a dedicated section in the documentation or (as is your patch) in the docstring for `parse`.

As I said, your patch looks good and definitely useful, so I'll close this issue.
