### 2021-07-22 jamesrhester/Lerche.jl#19 @ziotom78
Sorry, I'm reopening this because I realized that the docs you linked above refer to Lark and not to Lerch.jl. Is the improvement in the documentation already available somewhere? The last commit I see in `master` dates back to Jun 18.
