### 2021-07-22 #3497 @ziotom78
@jamesrhester , I am posting this here instead of opening an issue, because this is mostly relevant in the context of this review. I am not able to find any information about how to contribute to your code (see the point «**Community guidelines**» in the checklist and the [JOSS guide](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)).
