### 2021-07-22 jamesrhester/Lerche.jl#19 @erezsh
I will leave that to James to answer. But I believe Lerche's behavior should match Lark's in this case.
