### 2021-07-22 #3497 @ziotom78
@jamesrhester , I cannot find any performance claim of Lerche over Lark, and while this is not mandatory for a successful review at JOSS, it would make a valuable addition and would probably find its place in the *Statement of need* in the paper itself.

I created the simple JSON parser described in the Lerche.jl documentation and the same parser in Lark, and then I benchmarked the time required to open a ~500 KB file. Lerche.jl took ~0.3s, while Lark took 4s: Lerche.jl is an order of magnitude faster!

It would be nice to add a xstatement about this, of course after having validated these benchmarks.
