### 2021-07-23 c26641d
Improved documentation: - How to contribute - Document exceptions - How to raise issues
