+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Documentation","page":"API","title":"API Documentation","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"CurrentModule = Lerche","category":"page"},{"location":"api/#Parsing","page":"API","title":"Parsing","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"A parser is created from a Lark grammar by calling the Lark constructor. Parsing is initiated by calling parse.","category":"page"},{"location":"api/","page":"API","title":"API","text":"Lark(grammar::String;options...)\nLark(grammar::IOStream,source;options...)\nLerche.open(grammar_filename;rel_to=nothing,options...)\nLerche.parse(l::Lark,text;start=nothing,on_error=nothing)","category":"page"},{"location":"api/#Lerche.Lark-Tuple{String}","page":"API","title":"Lerche.Lark","text":"Lark(grammar::String;options...)\n\nCreate a Lark parser based on the Lark grammar expressed in grammar. Options are as follows:\n\nparser Decides which parser engine to use, \"earley\" or \"lalr\". (Default: \"lalr\")      Note: \"lalr\" requires a lexer. \"earley\" is not currently implemented.\nlexer Decides whether or not to use a lexer stage\n\"standard\": Use a standard lexer\n\"contextual\": Stronger lexer (only works with parser=\"lalr\")\n\"auto\" (default): Choose for me based on grammar and parser\ntransformer - Applies the transformer to every parse tree\ndebug - Affects verbosity (default: False)\nkeep_all_tokens - Don't automagically remove \"punctuation\" tokens (default: False)\npostlex - Lexer post-processing (Requires standard lexer. Default: None)\nstart - The start symbol (Default: start)\npropagate_positions - Propagates [line, column, end line, end column] attributes into all tree branches.\nlexer_callbacks - Dictionary of callbacks for the lexer. May alter tokens during lexing. Use with caution.\n\n\n\n\n\n","category":"method"},{"location":"api/#Lerche.Lark-Tuple{IOStream, Any}","page":"API","title":"Lerche.Lark","text":"Lark(grammar::IOStream,source;options...)\n\nCreate a Lark parser based on the Lark grammar found by reading grammar, whose source is recorded as source. See above for options.\n\n\n\n\n\n","category":"method"},{"location":"api/#Lerche.open-Tuple{Any}","page":"API","title":"Lerche.open","text":"open(grammar_filename;rel_to=nothing,options...)\n\nOpen grammar_filename and read in a Lark grammar, returning a Lark parser. If rel_to is not nothing, grammar_filename is relative to rel_to. Options are the same as for Lark(grammar::String;options...).\n\n\n\n\n\n","category":"method"},{"location":"api/#Lerche.parse-Tuple{Lark, Any}","page":"API","title":"Lerche.parse","text":"parse(l::Lark,text;start=nothing,on_error=nothing)\n\nParse text using parser l, returning a parse tree. If start is not nothing, it identifies the start symbol. on_error is currently ignored.\n\n\n\n\n\n","category":"method"},{"location":"api/#Working-with-the-parse-tree","page":"API","title":"Working with the parse tree","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"transform transforms the parse tree according to rules defined by the user using @rule and @inline_rule macros. Tokens will also be processed if visit_tokens returns true for that transformer type.  This will significantly slow down parse tree processing.","category":"page"},{"location":"api/","page":"API","title":"API","text":"@rule s\
[…]
