### 2021-07-23 jamesrhester/Lerche.jl#19 @jamesrhester
I have updated the `Lerche.jl` documentation in  c26641d111f87a710add563a1b68da5137f21c3d by adding to the docstring for `parse` (as for Lark), and including the exceptions in the API docs. There is also now a short paragraph in the top-level `README.md`.
