### 2021-07-23 jamesrhester/Lerche.jl#18 @jamesrhester
I have removed the stable docs badge. The stable docs are auto-generated by Documenter.jl each time a stable version number is reached (I think that will be 0.5.0). As this was not in place for 0.4.0, there are not yet any docs for a stable version.
