### 2021-07-23 #3497 @jamesrhester
> @jamesrhester , I am posting this here instead of opening an issue, because this is mostly relevant in the context of this review. I am not able to find any information about how to contribute to your code (see the point «**Community guidelines**» in the checklist and the [JOSS guide](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)).

I have just added two short sections "Issues" and "Contributions" to the top-level README. Hopefully these are sufficient.
