### 2021-07-23 #3497 @jamesrhester
> @jamesrhester , I cannot find any performance claim of Lerche over Lark, and while this is not mandatory for a successful review at JOSS, it would make a valuable addition and would probably find its place in the _Statement of need_ in the paper itself.
>
> I created the simple JSON parser described in the Lerche.jl documentation and the same parser in Lark, and then I benchmarked the time required to open a ~500 KB file. Lerche.jl took ~0.3s, while Lark took 4s: Lerche.jl is an order of magnitude faster!
>
> It would be nice to add a xstatement about this, of course after having validated these benchmarks.

An earlier draft contained some performance claims but we decided that on balance comparing to interpreted Python was not very helpful when things like PyPy can dramatically improve Python execution speed. I will consult with my coauthor on suitable additional text, as it is true that most people do expect some statements about execution speed in the context of Julia.
