### 2021-07-23 #3497 @ziotom78
> An earlier draft contained some performance claims but we decided that on balance comparing to interpreted Python was not very helpful when things like PyPy can dramatically improve Python execution speed. I will consult with my coauthor on suitable additional text, as it is true that most people do expect some statements about execution speed in the context of Julia.

I see. I just wanted to raise a flag because this is what people usually look after when they see a Python package ported to Julia, but  I do not think it's mandatory to include this information.
