### 2021-07-30 8dfff28
Fix construction of token_func call.
A massive time overhead was incurred when using token_func. This
has been largely removed.
