### 2021-07-30 342065a
Added @terminal macro for token processing (thanks to @peterahrens).
