### 2021-07-31 #3497 @ziotom78
>  @ziotom78, please update us on how your review is going (this is an automated reminder).

My review is almost finished; there was just the issue whether to include some statement about the performance or not, but if the authors don't feel to do this, that's fine with me.
