### 2021-07-31 #3497 @eschnett
I have reviewed the paper and the software. This paper is well written; it gives a brief overview of the software (the parser), but doesn't go into any details about its usage or functionality. I think this is fine, since the README and the documentation of the software do so. The software itself seems mature and has been used for a few years in a few projects. I hope that this publication will increase the reach of this Julia package – it could be useful for many other Julia packages as well.
