### 2021-07-31 #3497 @eschnett
A remark to the editors: If software is hosted on Github, then several of the items on the checklist can be checked off almost automatically. Furthermore, if the software is written in Julia, then quite a few more can be checked off almost automatically, since the Julia packaging standards and templates almost automatically lead to well-designed and easy-to-use software.
