### 2021-08-02 65452be
Added paragraph about performance.
