+++ paper.md
@@ -84,6 +84,22 @@ generated by `Lerche` from this EBNF. `Lerche` is also used by
 files written in the Crystallographic Information Framework (CIF)
 format [@Bernstein:aj5269].
 
+# Performance
+
+One of Julia's attractions is the potential for performance close to
+that of a compiled language, while retaining features normally found in
+high-level interpreted languages.  Fully realising this potential usually
+requires some investment in code design and profiling, which has so
+far been minimal for `Lerche.jl`. Nevertheless, parsing a 500K mmCIF
+text file from the worldwide Protein Data Bank [@wwpdb] with `Lerche.jl`
+is around 3 times faster than using Lark with CPython, if initial compilation time is
+ignored. This improvement essentially disappears if the PyPy
+just-in-time Python interpreter is used instead of CPython. Therefore, although further
+improvements to `Lerche.jl` runtime performance are likely possible and may
+be realised in future versions, at this point switching to Julia and 
+`Lerche.jl` to improve performance should be weighed carefully against
+the option of simply switching to a faster Python interpreter.
+
 # Features
 
 `Lerche` optionally augments the range of applicable EBNF grammars by
