### 2021-08-02 #3497 @jamesrhester
We have just added a paragraph about performance as it might be of interest.
