### 2021-08-02 #3497 @ziotom78
> We have just added a paragraph about performance as it might be of interest.

Thanks @jamesrhester , I have read the paragraph. I am fully satisfied, my review is complete.
