### 2021-08-02 #3497 @sbenthall
@jamesrhester I'm having trouble generating the PDF with the JOSS tools.
I wonder if there are any obvious discrepancies with the JOSS submission requirements on your end.
