### 2021-08-03 #3497 @jamesrhester
Please check a [few replies back on the Github issue](https://github.com/openjournals/joss-reviews/issues/3497#issuecomment-890628606) to see the correct form of the command when the paper is in a branch.
