### 2021-08-11 3e26f80
Version bump
In order to obtain a DOI on Zenodo re-release 0.5.0 as 0.5.1 (no other changes).
