### 2021-08-11 #3497 @jamesrhester
I have done so, the DOI for version 0.5.1 is 10.5281/zenodo.5178771
