### 2021-08-20 #3497 @kthyng
@jamesrhester Please update the metadata in your Zenodo archive so that the title and author list exactly match your JOSS paper.
