### 2021-08-23 #3497 @danielskatz
👋 @jamesrhester - note we just need you to update your zenodo metadata to proceed
