### 2021-08-23 #3497 @danielskatz
Additionally, in proofreading the paper, your bib entries for software packages should each have a title field with the title of the package.
