### 2021-08-23 9ab7787
small updates to paper.md
