+++ paper.md
@@ -31,7 +31,7 @@ grammars describe text that can be unambiguously parsed based
 only on the tokens already seen and the next token of input [@lalr].
 LALR(1) stands out for being able to parse most programming languages,
 while guaranteeing O(n) run-time complexity and very light memory use.
-Julia package `Lerche` automatically generates a parser that processes
+The `Lerche` Julia package automatically generates a parser that processes
 any data file or domain-specific language that can be described using
 a LALR(1) EBNF. The parse tree can be immediately transformed into
 application-specific data structures using user-supplied rules. This
@@ -44,8 +44,8 @@ Standards for scientific data formats are essential for correct
 interpretation of data produced and consumed by parties that are not
 in direct contact.  Standardisation is also becoming increasingly
 important as requirements for accessible and reusable data increase
-among funding bodies and publishers. Use of formal description
-languages, such as EBNF, remove ambiguity that may arise when using
+among funding bodies and publishers. The use of formal description
+languages, such as EBNF, removes ambiguity that may arise when using
 natural language descriptions for these standards. Furthermore, a
 parser that is machine-generated from an EBNF description is
 guaranteed to generate a correct parse tree from conformant data,
@@ -70,15 +70,15 @@ projects.
 Other native parser tools available for Julia projects include
 `ParserCombinator` [@parsercombinator], `Pegparser` [@pegparser], and
 the built-in Julia macro system. None of these take EBNF as input. Reliance
-on EBNF parser generators external to the Julia environment, for example
-by calling Lark or pre-generating parsers, complicate distribution and
-create a barrier to user contribution to packages.
+on EBNF parser generators external to the Julia environment, for example,
+by calling Lark or pre-generating parsers, complicates distribution and
+creates a barrier to user contribution to packages.
 
 dREL [@spadaccini_drel:_2012] is a relational data processing language
 for specifying crystallographic algorithms. The dREL LALR(1) EBNF [@drelenbf]
 was developed using `Lerche` to verify its correctness and conformance
-to LALR(1) requirements. Package
-`DrelTools` [@dreltools] is built around the parser automatically
+to LALR(1) requirements. The
+`DrelTools` [@dreltools] package is built around the parser automatically
 generated by `Lerche` from this EBNF. `Lerche` is also used by
 `CrystalInfoFramework` [@cifjl] to generate the parser for data
 files written in the Crystallographic Information Framework (CIF)
