### 2021-08-23 #3497 @danielskatz
Finally, I suggest a number of small changes in https://github.com/jamesrhester/Lerche.jl/pull/22
