### 2021-08-24 a2361ac
Add titles to bibliography entries.
