+++ paper.bib
@@ -18,6 +18,7 @@
 
 @misc{cifjl,
   author =       {J. R. Hester},
+  title  =       {Crystal{I}nfo{F}ramework.jl},
   url =          {https://github.com/jamesrhester/CrystalInfoFramework.jl},
   year =         2021,
   publisher =    {Github},
@@ -26,6 +27,7 @@
 
 @misc{drelenbf,
   author =       {J. R. Hester},
+  title =        {A Draft annotated grammar for d{REL}},
   year =         2021,
   journal =      {Github repository},
   publisher =    {Github},
@@ -35,6 +37,7 @@
 
 @misc{dreltools,
   author =       {J. R. Hester},
+  title =        {Drel{T}ools.jl},
   url =          {https://github.com/jamesrhester/DrelTools.jl},
   year =         2021,
   publisher =    {Github},
@@ -57,6 +60,7 @@
 
 @misc{lark,
   author =       {E. Shinan},
+  title =        {Lark},
   url =          {https://github.com/lark-parser/Lark},
   year =         2021,
   publisher =    {Github},
@@ -65,6 +69,7 @@
 
 @misc{parsercombinator,
   author =       {A. Cooke},
+  title =        {Parser{C}ombinator.jl},
   url =          {https://github.com/andrewcooke/ParserCombinator.jl},
   year =         2021,
   publisher =    {Github},
@@ -99,6 +104,7 @@
 }
 
 @misc{pegparser,
+  title = {{PEGP}arser.jl},
   url={https://github.com/abeschneider/PEGParser.jl},
   year={2020},
   author={A. Schneider},
