### 2021-08-24 #3497 @jamesrhester
I have updated the title and author list for the Zenodo metadata, merged Daniel's editorial fixes and added title fields where missing to the .bib file.
